Refactor (HomeViewModel): handle method
