+++ solutions/devsprint-bruno-faganello-2/SpaceApp/Utils/Extensions/Int+Extensions.swift
@@ -1,0 +1,10 @@
+//
+//  Int+Extensions.swift
+//  SpaceApp
+//
+//  Created by Paolo Prodossimo Lopes on 02/05/22.
+//
+
+extension Int {
+    var asLaunchCode: String { "#\(self)" }
+}
