+++ solutions/devsprint-bruno-faganello-2/SpaceApp/ViewModel/HomeViewModel.swift
@@ -12,7 +12,6 @@ final class HomeViewModel: ObservableObject {
     //MARK: - Propertie
     private let service: SpaceXAPIProtocol
     
-    @Published private(set) var launchs: [Launch] = []
     @Published private(set) var nextLaunch: NextLaunchViewData?
     
     //MARK: - Constructor
@@ -22,90 +21,20 @@ final class HomeViewModel: ObservableObject {
     
     //MARK: - Methods
     func start() {
-        fetchAllLaunch()
         fetchNextLaunch()
     }
     
     //MARK: - Helpers
-    private func fetchAllLaunch() {
-        Task { [weak self] in
-            guard let self = self else { return }
-            let allLaunchs = await service.fetchAllLaunches() ?? []
-            DispatchQueue.main.async { self.launchs = allLaunchs }
-        }
-    }
-    
     private func fetchNextLaunch() {
         Task { [weak self] in
             guard let self = self else { return }
             let nextLaunch = await service.fetchNextLaunch()
-            DispatchQueue.main.async {
-                guard let nl = nextLaunch else {
-                    self.nextLaunch = nil
-                    return
-                }
-                self.nextLaunch = .init(nextLaunchResponse: nl)
-            }
+            DispatchQueue.main.async { self.serviceResultHandler(nextLaunch) }
         }
     }
-}
-
-struct NextLaunchViewData: Identifiable {
     
-    private let nextLaunchResponse: NextLaunchResponseModel?
-    
-    init(nextLaunchResponse: NextLaunchResponseModel?) {
-        self.nextLaunchResponse = nextLaunchResponse
+    private func serviceResultHandler(_ response: NextLaunchResponseModel?) {
+        guard let nl = response else { self.nextLaunch = nil; return }
+        self.nextLaunch = .init(nextLaunchResponse: nl)
     }
-    
-    let id: UUID = .init()
-    
-    var rocketName: String {
-        nextLaunchResponse?.aircraftName ?? .emptyString
-    }
-    
-    var launchNumber: String {
-        let codeNumber = nextLaunchResponse?.flightNumber ?? .zero
-        return codeNumber.asLaunchCode
-    }
-    
-    var launchDescription: String {
-        nextLaunchResponse?.detail ?? .emptyString
-    }
-    
-    var launchDate: String {
-        let date = nextLaunchResponse?.launchDate ?? .emptyString
-        return date.dateFormattedFromUTC
-    }
-    
-    var iconImageURL: URL {
-        let pth = nextLaunchResponse?.links.patch.small ?? ""
-        return URL(string: pth)!
-    }
-    
-    var description: String {
-        nextLaunchResponse?.detail ?? .emptyString
-    }
-}
-
-extension String {
-    
-    static var emptyString: String { "" }
-    
-    var dateFormattedFromUTC: String {
-        let fromFormatter: DateFormatter = .init()
-        fromFormatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'"
-        fromFormatter.timeZone = .init(abbreviation: "UTC")
-        
-        let date = fromFormatter.date(from: self) ?? Date()
-        
-        let toFormatter: DateFormatter = .init()
-        toFormatter.dateFormat = "MMMM d, yyyy"
-        
-        return toFormatter.string(from: date)
-    }
-}
-
-extension Int {
-    var asLaunchCode: String { "#\(self)" }
 }
