+++ solutions/devsprint-bruno-faganello-2/SpaceAppTests/HomeViewModelTests.swift
@@ -8,6 +8,9 @@
 import XCTest
 @testable import SpaceApp
 
+import UIKit
+import SwiftUI
+
 final class HomeViewModelTests: XCTestCase {
     
     //MARK: - Properties
